Add removal of todos.
This means, rather than to handle a stream of items, we're concerned with a stream of operations that transform those items. A subtle shift that allow us to view our app from an even higher persective.

Intents are transformed into operations, which in turn transform the state of the app.

Following this gist for reference: http://blog.edanschwartz.com/2015/09/18/dead-simple-rxjs-todo-list/#requirement-2-a-user-can-remove-an-item-from-their-list