Add router @cycloidal/driver-router.
The router works with uniloc. Once initialized (using routes and alias definition), a history object will be created in driver-router. The observable sequence on the route will begin sending updates as soon as history has decoded the url.

Updates on the sequence are being sent for every observer that joins, even late. This is done with a hot observable.
As long as no route is present, the rendering of the view is skipped.

This way, we eliminate initialState from view().

npm install after update